Replaced admin login with member login.
The authentication page did not reflect the fact that memberships could
also log in and attend presentations. The page gave the impression that
membership login was just for administrator, which it was previously.

The authentication page now explains in more detail why there is two
login forms, which one to use and it provides useful links to ITKs pages
for registrering membership cards and resetting of membership passwords.