Changing participants lookup.
Since the member table exists on a different database, the previous code
would create a query selecting from the members table and joining from
the participations table. This obviously resultet in an error.

This code will select all member IDs and user IDs and call #find on the
respective classes by calling #constantize on participle_type.