Correcting a bug in a cucumber step.
The bug occured in a step of type "Then I should not [..]", which is
devised in a way such that it calls its respective "Then I should [..]"
twin step and expect it to throw an error.

However, it incorrectly just called itself and created an infinite
loop.  The error that actually was thrown was "stack level too deep
(SystemError)" and thus the step never actually failed.

Additionally, the fix has a positive effect on run time.

  $ time cucumber
  real    0m23.430s
  user    0m22.937s
  sys     0m0.427s

  $ time cucumber
  real    0m19.697s
  user    0m19.153s
  sys     0m0.463s